changes order statuses

+++ src/main/kotlin/ru/foodbox/delivery/controllers/auth/body/CreateCartRequestBody.kt
@@ -2,12 +2,13 @@ package ru.foodbox.delivery.controllers.auth.body
 
 import ru.foodbox.delivery.data.DeliveryType
 import ru.foodbox.delivery.services.dto.AddressDto
+import java.math.BigDecimal
 
 data class CreateCartRequestBody(
     val deviceId: String,
     val deliveryType: DeliveryType,
     val deliveryAddress: AddressDto?,
     val departmentId: Int,
-    val deliveryPrice: Double,
-    val freeDeliveryPrice: Double?
+    val deliveryPrice: BigDecimal,
+    val freeDeliveryPrice: BigDecimal?
 )
